Pass the Python executable found by UsePythonVersion to CMake.
In Windows, CMake FindPythonInterpr uses the registry to find Python if none if provided. Instead of leaving the decision to chance of which Python has written last into the registry, pass the Python found by UsePythonVersion into CMake.

To allow both Unix and Windows, and pipelines that finds an specific Python or not, the template parameter is optional and expands or not into a CMake parameter. This should keep the working pipelines the same, while modifying the failing pipeline to work.