toolchain: template and homogenise
This begins the homogenisation of the toolchain builds.  By pushing all
the builds into a template, it becomes more feasible to make the
toolchain job a job template.  At that point, it will become possible to
synchronise the VS2017, VS2019 builds across development and stable,
reducing the pipelines.  This also migrates the builds to
multi-repository checkouts, full checkouts, and pipeline artifacts.  The
fallout of this will ripple through the rest of the builds, but, will
enable better build strategies.