Adding a more details to the getting starting doc
I had to google the locations of the UniversalCRTSdkDir as an example of where we could make the docs a little easier for those not familair with the toolchain.

I've added a couple of edits to make it a tiny bit easier for those who don't hae a wealth of experiance with Windows dev.