Installer: clean up the custom action a bit
Remove the now extraneous logic as there is a better understanding of
the scheduling sequence and handling of the cleanup.  The last item that
remains to be resolved is the committing of the alterations to the
database.  We know now that there will be a single action executed
deferred, not impersonated, after InstallExecute to ensure that we can
deploy the auxiliary support files.  The elevate privileges should allow
us to also mutate the on-disk database for the cached MSI to allow
tracking of the deployed files so that we can clean them up on the way
out.  The clever idea of using the `RemoveFile` table allows us to avoid
using a custom action for the removal process.