Update swift-format.yml
Apply a variety of changes to the swift-format build.  We now use the swift-installer-scripts repository to package the code formatter into a distributable MSI.  Update the product version tag in the package.

Hoist the checkout steps to the initial operations rather than delaying the location as this was done to workaround the `swift build` invocation.  Instead we now explicitly change to the appropriate location prior to building.