build: use DWARF for the SDK builds
This adjusts the builds for the SDK to switch to DWARF.  This is
problematic in many ways:
  - it adds code bloat
  - it slows down linking
  - it slightly reduces incremental builds
  - the debug information is not possible to be saved
  - the resultant binaries are in violation of the COFF specification

Given all the issues, this is still valuable as a change to enable
easier debugging.  The official builds remain building with CodeView
which means that the PDBs can be linked in parallel, we do not bloat the
released binaries, while we get a better debugging experience locally.