build: shuffle some state construction
Create a new global variable to compute the host runtime directory.
This is required to execute tools that are built for the toolchain.
The long term goal here is to build a small subset of tools using a
build of SPM during the toolchain build to enable packaging the set
into the toolchain distribution.