build: add DocC to the default built
DocC is not yet part of the distribution on Windows as we are
progressing towards passing the test suite.  However, other products
were released before the test suites were completely passing and that
would remain an option here.  Begin building this tool as part of the
build as it is meant to be part of the distribution.