GHA: wire up Foundation macros to swift-corelibs-foundation
We are already building the Foundation macros but did not pass that
along to the Foundation build. This is now required and failure to pass
it along results in a build failure. Wire it up so that we can build and
avoid another build of the macros.

(cherry picked from commit 04633887a6a9ac44e0ae185290aca8dfc2b1d475)