Refactor matrix definitions
The various build matrix definitions used to be duplicated across
multiple jobs. This changes the matrix definitions to have a single
definition for every OS-Arch target, removing duplication. In addition,
this change makes it easier to set up a separate arm64 Windows build.

This PR breaks the Mac build. It will be fixed as all jobs will be
updated to use the new setup. Subsequent PRs will introduce Windows-only
and Android-only matrices passed to the sub-build.

The end goal is to have a single matrix definition for every platform
and remove many of the inputs to the `swift-toolchain` sub-build. This
also allows for re-merging the Mac and Windows sub-builds, reducing
build complexity.