[6.1] Temporarily specify Windows SDK and MSVC versions
This is a Cherry-Pick of #913

This is necessary to work around an issue with the latest version
(26100) of the Windows SDK where changes in headers result in circular
module dependencies.
For the compilers build, this installs and uses MSVC 14.42 because the
current build does not work with MSVC 14.43. In addition, this removes
the 26100 Windows SDK to get the compilers to build until the bootstrap
toolchain supports the new SDK.

See #909 for details.